Extract OneKey public library in a dedicated file and document all public methods
