+++ paf-mvp-frontend/CONTRIBUTE.md
@@ -1,0 +1,15 @@
+# OneKey: contribute to the frontend project
+
+You can run `npm start` to create a bundle and start server. The bundle will be available on
+http://localhost:3000/dist/app.bundle.js
+
+You can open http://localhost:3000/ to see how it looks like in the sandbox. To test app in "Real Environment", you can
+inject script (as described above) to any website.
+
+There are also available scripts:
+
+* `npm run lint` and `npm run lint:prettier` - to lint a code style.
+
+## Deployment
+
+Once deployed, the bundle will be available by the path: `https://{{domain}}/app.bundle.js`
