+++ paf-mvp-frontend/src/global.d.ts
@@ -1,5 +1,5 @@
 import { IProcessingQueue } from './utils/queue';
-import { IOneKeyLib } from '@onekey/frontend/lib/paf-lib';
+import { IOneKeyLib } from '@onekey/frontend/lib/i-one-key-lib';
 
 export type Window = WindowProxy &
   typeof globalThis & {
